Move common code to utils
