+++ loader-tests/utils.ts
@@ -1,7 +1,16 @@
+import './dom-parser.js'
+
 import {
+  createDownloadTask,
   enableTestTime,
+  getLoaderForText,
   getTestEnvironment,
+  loaders,
+  type PreviewCandidate,
+  previewCandidates,
+  previewCandidatesLoading,
   setBaseTestRoute,
+  setPreviewUrl,
   setRequestMethod,
   setupEnvironment,
   userId
@@ -10,6 +19,12 @@ import type { ReadableAtom } from 'nanostores'
 import { readFile } from 'node:fs/promises'
 import { isAbsolute, join } from 'node:path'
 import { styleText } from 'node:util'
+
+export interface Feed {
+  htmlUrl: string
+  title: string
+  url: string
+}
 
 export async function readText(path: string): Promise<string> {
   let absolute = path
@@ -111,4 +126,60 @@ export function success(msg: string, details?: string): void {
     msg += ` ${styleText('gray', details)}`
   }
   print(styleText('green', styleText('bold', '✓ ') + msg))
+}
+
+export async function fetchAndParsePosts(url: string): Promise<void> {
+  try {
+    let task = createDownloadTask()
+    let textResponse = await task.text(url)
+    let candidate: false | PreviewCandidate = getLoaderForText(textResponse)
+    if (!candidate) {
+      error(`Can not found loader for feed ${url}`)
+      return
+    }
+    let loader = loaders[candidate.loader]
+    let { list } = loader.getPosts(task, url, textResponse).get()
+    if (list.length === 0) {
+      error(`Can not found posts for feed ${url}`)
+      return
+    }
+    success(url, list.length + (list.length > 1 ? ' posts' : ' post'))
+  } catch (e) {
+    error(e, `During loading posts for ${url}`)
+  }
+}
+
+export async function findRSSfromHome(feed: Feed): Promise<void> {
+  let unbindPreview = previewCandidates.listen(() => {})
+  try {
+    setPreviewUrl(feed.htmlUrl)
+    await timeout(10_000, waitFor(previewCandidatesLoading, false))
+    if (previewCandidates.get().some(c => c.url === feed.url)) {
+      success(`Feed ${feed.title} has feed URL at home`)
+    } else if (previewCandidates.get().length === 0) {
+      error(
+        `Can’t find any feed from home URL or ${feed.title}`,
+        `Home URL: ${feed.htmlUrl}\nFeed URL: ${feed.url}`
+      )
+    } else {
+      error(
+        `Can’t find ${feed.title} feed from home URL`,
+        `Home URL: ${feed.htmlUrl}\n` +
+          `Found: ${previewCandidates
+            .get()
+            .map(i => i.url)
+            .join('\n       ')}\n` +
+          `Feed URL: ${feed.url}`
+      )
+    }
+  } catch (e) {
+    error(
+      e,
+      `During searching for feed from home URL\n` +
+        `Home URL: ${feed.htmlUrl}\n` +
+        `Feed URL: ${feed.url}`
+    )
+  } finally {
+    unbindPreview()
+  }
 }
